Monday, 2nd June, 2003, Al Riddoch <alriddoch@zepler.org> 	* Move to requiring sigc++ 1.2. 	* Move to pkgconfig completely. 	* varconf/config.cpp, varconf/variable.cpp: Fix a couple of warnings. 	* Increment version to 0.6.0, interface version 1.0.0 and release.

+++ varconf/config.cpp
@@ -189,7 +189,7 @@ void Config::getCmdline( int argc, char** argv)
   bool fnd_sec = false, fnd_nam = false;
   size_t mark = 2;
 
-  for ( size_t i = 1; i < argc; i++) {
+  for ( size_t i = 1; i < (size_t)argc; i++) {
     if ( argv[i][0] == '-' && argv[i][1] == '-' && argv[i][2] != '\0') {
       arg = argv[i];
       fnd_sec = fnd_nam = false;
